pyproject.toml
[project]
name = "repo-sanitizer"
version = "2.0.0"
requires-python = ">=3.8"
dependencies = ["GitPython", "Typer", "Rich", "InquirerPy", "PyYAML", "requests"]

[project.scripts]
clean-repo = "repo_sanitizer.main:app"
